Update project to support Java 8
https://eaflood.atlassian.net/browse/WC-273

As part of work to move the service to MongoDb we have found we need to move our main waste carriers service to Java JDK 8.

To remove the complexity of having multiple JDK's in the service we are ensuring all the Java elements of the service are updated to support building and running with OpenJDK 8.

WIP