Update README with missing info
This change adds to missing details to the README.

The first is to explain how the project is now using Maven Wrapper to ensure a version of Maven is available for building the project.

The second is bringing the app inline with anothr service, and defaulting Maven not to run tests when it builds. It also covers the reasons why, and how to run them manually.