Incorporate env vars into the configuration.yml
https://eaflood.atlassian.net/browse/WC-159

To improve how the service can be deployed rather than have to hard code the OS Places API key into a config file (and then have to store that config file somewhere) we've taken advantage of the upgrade in the Dropwizard framework to use environment variable substitution.

Now when the app is deployed you must provide the OS Places key in an env var called WCRS_OSPLACES_KEY.The log files will default to where they would go on our vagrant build, but this can be overridden.