Remove skip test in pom.xml & use cmd arg instead
Because the project requires an actual connection to OS Places in order to run the build, this can make it a problem in CI and Jenkins jobs where that access may not be available.

So to prevent issues when building we added a skip test flag to the pom.xml, however we then found we could not get `mvn test` to run either.

Plus expected behaviour is that the tests would automatically run hence we've decided to update the project to remove the flag and instead use a command line option to govern when whether tests should be run as part of a build.

This makes it more explicit and ensures the project behaves as expected.