Add ability to mock calls to OS Places
https://eaflood.atlassian.net/browse/WC-366

Previously on other services we've had a mock for the address lookup so we aren't hitting osplaces with thousands of requests during performance tests.

On waste carriers however it use this service which returns results in a different format and so isn;t suitable.

This change updates the service to include the ability to 'switch-on' mocking calls to OS Places. So when enabled rather than quering OS places for a result, it instead will read from a file that contains a json copy of a previous result.

As far as any external client is concerned, the call has been made as normal.

WIP