Added OSPlacesMock and refactored resources
This changes adds the OSPlacesMock class which in an implementation of IOSPlaces that does not call the actual OS Places API.

It also refactors the resources to bring together the search by Uprn and search by postcode into a single `SearchResource`. It also deletes what appears to be an unused resources to get the status of OS Places.