Fix issue reading mock resource when a JAR
This changes resolves an issue only found when the project was compiled as a jar. Once a JAR the resources files are truly part of the file system hence you can no longer access them using classes such as `File`.

This post gives an explanation of the reason why https://stackoverflow.com/a/27390373/6117745

In short, the only way to access a resource file once the project is compiled as a JAR seems to be as a stream, hence the change in implementing of the `OSPlacesMock` class.