Add ability to mock calls to OS Places (#17)
https://eaflood.atlassian.net/browse/WC-366

Previously on other services we've had a mock for the address lookup so we aren't hitting osplaces with thousands of requests during performance tests.

On waste carriers however it use this service which returns results in a different format and so isn;t suitable.

This change updates the service to include the ability to 'switch-on' mocking calls to OS Places. So when enabled rather than quering OS places for a result, it instead will read from a file that contains a json copy of a previous result.

As far as any external client is concerned, the call has been made as normal.

WIP

* Added OSPlacesMock and refactored resources

This changes adds the OSPlacesMock class which in an implementation of IOSPlaces that does not call the actual OS Places API.

It also refactors the resources to bring together the search by Uprn and search by postcode into a single `SearchResource`. It also deletes what appears to be an unused resources to get the status of OS Places.

* Refactor Address to remove redudant get/sets

The address class is massive however that is down purely to the number of properties, and therefore getters and setters it has.

This change removes them, simplifying the class.

* General code tidy and reformat

Just moving things like braces to be consistent, and removing noise such as the `Assert.` in `Assert.assertFoo`.

* Fix issue reading mock resource when a JAR

This changes resolves an issue only found when the project was compiled as a jar. Once a JAR the resources files are truly part of the file system hence you can no longer access them using classes such as `File`.

This post gives an explanation of the reason why https://stackoverflow.com/a/27390373/6117745

In short, the only way to access a resource file once the project is compiled as a JAR seems to be as a stream, hence the change in implementing of the `OSPlacesMock` class.