Add support for test coverage in CI
This project has tests, and is hooked up to Travis-CI. However, it has never produced details on the level of test coverage.

With the move to hook all our projects to [SonarCloud](https://sonarcloud.io/organizations/defra/projects) we have decided to take this opportunity to add support to generate test coverage, so we can display it in SonarCloud.