Add do not use notice to README
We have been asked a few times about using this project for address lookup. However, it has been superseded by projects like the EA Addressbase Facade so it really should not be used by other services in Defra (though anyone else is free to check it out).

This adds a notice to try and make that info clear and up front.